requirements.txt
requests==2.31.0
pandas==2.1.4
beautifulsoup4==4.12.2
openpyxl==3.1.2
python-dotenv==1.0.0
tqdm==4.66.1
